feat: add analytics script
