+++ plugins/gatsby-remark-header-custom-ids/gatsby-ssr.js
@@ -62,16 +62,6 @@ exports.onRenderBody = ({setHeadComponents}, pluginOptions) => {
     })
   `;
 
-  const baidu_script = `
-    var _hmt = _hmt || [];
-    (function() {
-      var hm = document.createElement("script");
-      hm.src = "https://hm.baidu.com/hm.js?6c42f175403a43b961ef0799f22f1e69";
-      var s = document.getElementsByTagName("script")[0]; 
-      s.parentNode.insertBefore(hm, s);
-    })();
-  `;
-
   const style = icon ? (
     <style key="gatsby-remark-header-custom-ids-style" type="text/css">
       {styles}
@@ -83,10 +73,6 @@ exports.onRenderBody = ({setHeadComponents}, pluginOptions) => {
     <script
       key="gatsby-remark-header-custom-ids-script"
       dangerouslySetInnerHTML={{__html: script}}
-    />,
-    <script
-      key="gatsby-remark-header-custom-ids-script"
-      dangerouslySetInnerHTML={{__html: baidu_script}}
     />,
   ]);
 };
